added checkout page

+++ db/migrate/20211119041303_create_books.rb
@@ -3,9 +3,9 @@ class CreateBooks < ActiveRecord::Migration[6.1]
     create_table :books do |t|
       t.string :name
       t.string :author
-      t.decimal :user_rating
+      t.integer :user_rating
       t.string :review
-      t.decimal :price
+      t.integer :price
       t.integer :year
       t.text :description
       t.string :image_url
